requirements.txt
Click==7.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.1.0
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.10.3
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
visitor==0.1.3
Werkzeug==0.16.0
whitenoise==4.1.4
